add PackageController whit functionality and views

+++ Panda-Asp.Net-App/Panda.Models/BindingModels/CreatePackageInputModel.cs
@@ -8,7 +8,7 @@ namespace Panda.Models.BindingModels
     {
         public string Description { get; set; }
 
-        public decimal Weight { get; set; }
+        public double Weight { get; set; }
 
         public string ShippingAddress { get; set; }
